Fix mining

+++ crates/types/src/block_production.rs
@@ -6,7 +6,7 @@ use crate::H256;
 #[rtype(result = "()")]
 pub struct SolutionContext {
     pub partition_id: u64,
-    pub chunk_index: u64,
+    pub chunk_index: u32,
     pub mining_address: H256,
 }
